requirements.txt
django
django-environ
psycopg2-binary
pgvector
openai
djangorestframework
gunicorn
google-cloud-aiplatform
google-genai
langchain-google-vertexai
twilio
python-json-logger